Fixing double extraction from Qt widgets to use a float extraction instead

+++ Source/StatsGenerator/Delegates/SGMDFItemDelegate.h
@@ -129,7 +129,7 @@ class SGMDFItemDelegate : public QStyledItemDelegate
         QLineEdit* lineEdit = qobject_cast<QLineEdit* > (editor);
         Q_ASSERT(lineEdit);
         bool ok = false;
-        double v = lineEdit->text().toDouble(&ok);
+        double v = lineEdit->text().toFloat(&ok);
         model->setData(index, v);
       }
       else if (col == SGMDFTableModel::Axis)
